refactor: move EmptyFilter to components
