+++ src/components/EmptyFilter/i18n/index.ts
@@ -1,4 +1,4 @@
-import {registerKeysets} from '../../../../utils/i18n';
+import {registerKeysets} from '../../../utils/i18n';
 
 import en from './en.json';
 import ru from './ru.json';
