+++ src/containers/Storage/PaginatedStorageGroupsTable/StorageGroupsEmptyDataMessage.tsx
@@ -1,6 +1,6 @@
+import {EmptyFilter} from '../../../components/EmptyFilter/EmptyFilter';
 import {VISIBLE_ENTITIES} from '../../../store/reducers/storage/constants';
 import type {VisibleEntities} from '../../../store/reducers/storage/types';
-import {EmptyFilter} from '../EmptyFilter/EmptyFilter';
 
 import i18n from './i18n';
