+++ src/containers/Storage/PaginatedStorageNodesTable/StorageNodesEmptyDataMessage.tsx
@@ -1,7 +1,7 @@
+import {EmptyFilter} from '../../../components/EmptyFilter/EmptyFilter';
 import {VISIBLE_ENTITIES} from '../../../store/reducers/storage/constants';
 import type {VisibleEntities} from '../../../store/reducers/storage/types';
 import {NodesUptimeFilterValues} from '../../../utils/nodes';
-import {EmptyFilter} from '../EmptyFilter/EmptyFilter';
 
 import i18n from './i18n';
